Crate vqf

This crate reimplements most of the relevant parts of the VQF algorithm from https://github.com/dlaidig/vqf/blob/f2a63375604e0b025048d181ba6a204e96ce2559/vqf/pyvqf.py Currently this is just copy-pasted from the python code, but it should be made more idiomatic before actually using it. I have marked areas most likely to contain bugs with à¶

The original code is licensed under the MIT license, so this crate is also licensed under the MIT license.

Structs

VQFCoefficients

Vqf

VqfParameters

VqfState

Constants

EPS ğŸ”’

Functions

filterInitialState ğŸ”’

filterStep ğŸ”’

filterVec ğŸ”’

Type Definitions

Mat2x2 ğŸ”’

Mat2x3 ğŸ”’

Mat2x9 ğŸ”’

Mat3x3 ğŸ”’

Quat ğŸ”’

Vec2 ğŸ”’

Vec3 ğŸ”’